The common source file is divided into several
